Added utility for wrapping user data retrieved from Moria.
An object can be created by supplying a set of attributes, retrieved from Moria, to the constructor. The class contains some basic functionality for accessing the data. The API is created without having any specific requests, so we'll probably have to change and extend the API when we know how the users would like to use Moria. Anyway, its a start...